# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Photo Booth Production API",
    "description": "Production-ready REST API for Photo Booth SaaS with Session Lifecycle, Image Validation, and Secure Public Access.",
    "version": "1.1.0",
    "contact": {
      "name": "API Support",
      "email": "support@photobooth.com"
    }
  },
  "host": "backend-photobooth-production-e134.up.railway.app",
  "basePath": "/api/v1",
  "schemes": ["https", "http"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format: Bearer {token}"
    }
  },
  "paths": {
    "/sessions": {
      "post": {
        "tags": ["Sessions"],
        "summary": "Create a new photobooth session",
        "description": "Initializes a session in DRAFT then transitions to CAPTURING.",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["template_id", "layout_count"],
              "properties": {
                "event_name": {"type": "string"},
                "event_type": {"type": "string"},
                "location": {"type": "string"},
                "template_id": {"type": "integer"},
                "layout_count": {"type": "integer"},
                "duration": {"type": "integer", "description": "Duration in hours"}
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Session created",
            "schema": {"$ref": "#/definitions/Session"}
          }
        }
      }
    },
    "/sessions/{session_id}": {
      "get": {
        "tags": ["Sessions"],
        "summary": "Get session details",
        "parameters": [
          {"in": "path", "name": "session_id", "required": true, "type": "string"},
          {"in": "query", "name": "token", "type": "string", "description": "Secure token for public access"}
        ],
        "responses": {
          "200": {
            "description": "Session details",
            "schema": {"$ref": "#/definitions/Session"}
          },
          "401": {"description": "Invalid token"},
          "404": {"description": "Session not found"}
        }
      }
    },
    "/photos/upload": {
      "post": {
        "tags": ["Photos"],
        "summary": "Upload photo to session",
        "description": "Supports both authenticated and public sessions via token.",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {"in": "formData", "name": "photo", "type": "file", "required": true},
          {"in": "formData", "name": "session_id", "type": "string", "required": true},
          {"in": "formData", "name": "session_token", "type": "string"},
          {"in": "formData", "name": "filter", "type": "string"}
        ],
        "responses": {
          "201": {"description": "Photo uploaded and queued for processing"},
          "400": {"description": "Invalid image or session state"}
        }
      }
    },
    "/photos/strip-public": {
      "post": {
        "tags": ["Photos"],
        "summary": "Upload base64 photo strip",
        "description": "Used by frontend to upload generated strips anonymously.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["image_base64"],
              "properties": {
                "image_base64": {"type": "string"},
                "template_id": {"type": "integer"},
                "session_id": {"type": "string"},
                "session_token": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "Strip uploaded successfully"}
        }
      }
    }
  },
  "definitions": {
    "Session": {
      "type": "object",
      "properties": {
        "session_id": {"type": "string"},
        "token": {"type": "string"},
        "status": {"type": "string", "enum": ["DRAFT", "CAPTURING", "PROCESSING", "WAITING_PAYMENT", "PAID", "COMPLETED", "EXPIRED", "CANCELLED"]},
        "photo_count": {"type": "integer"},
        "expires_at": {"type": "string", "format": "date-time"},
        "photos": {
          "type": "array",
          "items": {"$ref": "#/definitions/Photo"}
        }
      }
    },
    "Photo": {
      "type": "object",
      "properties": {
        "id": {"type": "integer"},
        "original_url": {"type": "string"},
        "processed_url": {"type": "string"},
        "status": {"type": "string"},
        "is_anonymous": {"type": "boolean"}
      }
    }
  }
}